w3c-reports/12092014/gtk/runned-tests/media-source/SourceBuffer-abort-updating-expected.txt


FAIL SourceBuffer#abort() (video/webm; codecs="vorbis,vp8") : Check the algorithm when the updating attribute is true. assert_equals: expected (number) 0 but got (undefined) undefined(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/media-source/SourceBuffer-abort-updating.html:63:38
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/media-source/SourceBuffer-abort-updating.html:60:27)
FAIL SourceBuffer#abort() (video/mp4) : Check the algorithm when the updating attribute is true. assert_unreached: Browser doesn't support the MIME used in this test: video/mp4 Reached unreachable code(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_unreached@http://localhost:8800/resources/testharness.js:692:16
http://localhost:8800/media-source/SourceBuffer-abort-updating.html:34:29
step@http://localhost:8800/resources/testharness.js:798:30
isTypeSupported@http://localhost:8800/media-source/SourceBuffer-abort-updating.html:33:18
isTypeSupported@[native code]
http://localhost:8800/media-source/SourceBuffer-abort-updating.html:43:36
step@http://localhost:8800/resources/testharness.js:798:30
async_test@http://localhost:8800/resources/testharness.js:90:26
mediaTest@http://localhost:8800/media-source/SourceBuffer-abort-updating.html:42:15
global code@http://localhost:8800/media-source/SourceBuffer-abort-updating.html:95:14)
